Trying to bail out if build fails

+++ docker-world.sh
@@ -18,10 +18,10 @@ trap "rm $PWD/running" EXIT
 TARGET=${1:-full}
 
 # throw away building artefacts
-git stash
-git stash drop
+#git stash
+#git stash drop
 # get latest version
-git pull
+#git pull
 
 # only languages which have translations in transifex
 : ${langs:=en ca da de es fa fi fr gl hu id it ja km_KH ko lt nl pl pt_BR pt_PT ro tr ru uk zh_CN zh_TW}
@@ -31,10 +31,16 @@ git pull
 
 for l in $langs
   do
-    time /bin/bash ./docker-run.sh $TARGET LANG=$l
-    #time rsync -hvrzc -e ssh --progress output/html/$l qgis.osgeo.osuosl.org:/var/www/qgisdata/QGIS-Website/live/html
-    # 1 dec 2017: sync to qgis2
-    time rsync -hvrzc --delete --progress output/html/$l /var/www/qgisdata/QGIS-Website/live/html
+    ./docker-run.sh $TARGET LANG=$l
+    build_ok=$?
+    if [[ "$build_ok" = "0" ]]; then
+      echo "Build OK: syncing to web"
+      #time rsync -hvrzc -e ssh --progress output/html/$l qgis.osgeo.osuosl.org:/var/www/qgisdata/QGIS-Website/live/html
+      # 1 dec 2017: sync to qgis2
+      time rsync -hvrzc --delete --progress output/html/$l /var/www/qgisdata/QGIS-Website/live/html
+    else
+      echo "Build FAILED: not syncing to web";
+    fi
   done
 
 time rsync -hvzc -e ssh --progress output/html/version.txt /var/www/qgisdata/QGIS-Website/live/html
